refactoring - add qr code scanner page
