+++ MicroSocial/src/app/app-routing.module.ts
@@ -4,7 +4,7 @@ import { PreferencesComponent } from './preferences/preferences.component';
 import { AppComponent } from './app.component';
 import { SignInComponent } from './sign-in/sign-in.component';
 import { SignUpComponent } from './sign-up/sign-up.component';
-import { QrScannerComponent } from './qrScanner/qr-scanner/qr-scanner.component';
+import { QrScannerComponent } from './qr-scanner/qr-scanner.component';
 
 const routes: Routes = [
   { path: '', component: AppComponent },
