+++ MicroSocial/src/app/qr-scanner/qr-scanner.component.spec.ts
