+++ MicroSocial/src/app/qr-scanner/qr-scanner.component.ts
